+++ app/src/main/java/com/kc/shiptransport/mvp/acceptance/AcceptanceContract.java
@@ -24,7 +24,7 @@ public interface AcceptanceContract {
         // 显示选中周的任务计划
         void showWeekTask(List<String> dates, List<WeekTask> weekLists);
         // 每天任务量统计
-        void showDayCount(Double[] integers);
+        void showDayCount(Integer[] integers);
         // 是否显示加载
         void showLoading(boolean active);
         // 显示失败
