+++ app/src/main/java/com/kc/shiptransport/mvp/acceptance/AcceptancePresenter.java
@@ -167,14 +167,14 @@ public class AcceptancePresenter implements AcceptanceContract.Presenter {
                 .getDayCount()
                 .subscribeOn(Schedulers.io())
                 .observeOn(AndroidSchedulers.mainThread())
-                .subscribe(new Observer<Double[]>() {
+                .subscribe(new Observer<Integer[]>() {
                     @Override
                     public void onSubscribe(Disposable d) {
                         compositeDisposable.add(d);
                     }
 
                     @Override
-                    public void onNext(Double[] value) {
+                    public void onNext(Integer[] value) {
                         // 计算总计划量
                         view.showDayCount(value);
                     }
