+++ app/src/main/java/com/kc/shiptransport/mvp/amount/AmountPresenter.java
@@ -155,14 +155,14 @@ public class AmountPresenter implements AmountContract.Presenter{
                 .getDayCount(SettingUtil.TYPE_AMOUNT)
                 .subscribeOn(Schedulers.io())
                 .observeOn(AndroidSchedulers.mainThread())
-                .subscribe(new Observer<Double[]>() {
+                .subscribe(new Observer<Integer[]>() {
                     @Override
                     public void onSubscribe(Disposable d) {
                         compositeDisposable.add(d);
                     }
 
                     @Override
-                    public void onNext(Double[] value) {
+                    public void onNext(Integer[] value) {
                         // 计算总计划量
                         view.showDayCount(value);
                     }
