+++ app/src/main/java/com/kc/shiptransport/mvp/plan/PlanContract.java
@@ -18,7 +18,7 @@ public interface PlanContract {
         // 当前日期
         void showCurrentDate(String date);
         // 任务量
-        void showTaskVolume(Float value);
+        void showTaskVolume(Integer value);
         // 任务要求
         void showTaskRequire();
         // 当前任务总量
@@ -26,11 +26,11 @@ public interface PlanContract {
         // 本周任务分配
         void showWeekTask(List<String> dates, List<WeekTask> weekLists);
         // 每天任务量统计
-        void showDayCount(Double[] integers);
+        void showDayCount(Integer[] integers);
         void showLoading(boolean active);
         void showError(String msg);
         // 每日需求
-        void showDemandDayCount(Double[] datas);
+        void showDemandDayCount(Integer[] datas);
     }
 
     interface Presenter extends BasePresenter {
@@ -42,7 +42,7 @@ public interface PlanContract {
         // 任务要求
         void getTaskRequire();
         // 当前任务总量
-        void getTotalTaskVolume(Double[] integers);
+        void getTotalTaskVolume(Integer[] integers);
         // 本周任务分配
         void getWeekTask(int jumpWeek);
         // 每天任务量统计
